[actions,gaming_app] move weather magic action creation to action factory

+++ src/applications/gaming_app/WeatherListView.cpp
@@ -104,7 +104,7 @@ void WeatherListView::DeleteItem()
 // Name: WeatherListView::PopTrashedWeather
 // Created: ABR 2011-11-28
 // -----------------------------------------------------------------------------
-unsigned long WeatherListView::PopTrashedWeather()
+unsigned int WeatherListView::PopTrashedWeather()
 {
     if( trashedWeather_.empty() )
         return 0;
